+++ docs/changelog.md
@@ -2,6 +2,7 @@
 - Added a goodbye message when removing the bot from a server.
 - Fixed the bot not having proper command context.
 - Made the nickname changing option for transformations be false by default.
+- Updated Terms of Service and Privacy Policy links.
 
 ---